orionlx-plus: implement boot counter with rollback
After an update is performed, but before the boot pointer is changed a
a 1 byte boot counter value is stored at offset of 1MiB (block 0x800)
of the flash disk. The NAND flash erase block size is 1MiB and nothing
else is stored in this 1 MiB region.

The boot counter is disabled with values of 0 or 0xff.
If the value is between 2 and 0xfe, it is decremented and the boot slot
indicated by the boot pointer is used (presence of altboot.txt).
If the value is 1, the boot counter is left unchanged but the logical
complement of the boot pointer is used.

Mantis ID: 4736

Signed-off-by: George McCollister <george.mccollister@gmail.com>